WIP: test_connect_as_sink
BUG=b:149772936
BRANCH=none
TEST=this is a test

+++ common/usbc/usbc_task.c
@@ -77,7 +77,7 @@ void schedule_deferred_pd_interrupt(const int port)
  */
 void pd_interrupt_handler_task(void *p)
 {
-	const int port = (int) p;
+	const int port = (int) ((intptr_t) p);
 	const int port_mask = (PD_STATUS_TCPC_ALERT_0 << port);
 
 	ASSERT(port >= 0 && port < CONFIG_USB_PD_PORT_MAX_COUNT);
